@@ -1,3 +1,6 @@
-module.exports = {
-	plugins: ['tailwindcss', 'autoprefixer'],
+export default {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
 };
